# openapi.json
{
  "components": {
    "schemas": {
...
      "EventCreate": {
        "properties": {
          "chapter_id": {
            "title": "Chapter Id",
            "type": "string"
          },
          "ends_at": {
            "format": "date-time",
            "title": "Ends At",
            "type": "string"
          },
          "feedback_enabled": {
            "default": true,
            "title": "Feedback Enabled",
            "type": "boolean"
          },
          "help_options": {
            "items": {
              "type": "string"
            },
            "title": "Help Options",
            "type": "array"
          },
          "latitude": {
            "anyOf": [
              {
                "maximum": 90.0,
                "minimum": -90.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "locale": {
            "default": "nl",
            "enum": [
              "nl",
              "en"
            ],
            "title": "Locale",
            "type": "string"
          },
          "location": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Location",
            "type": "string"
          },
          "longitude": {
            "anyOf": [
              {
                "maximum": 180.0,
                "minimum": -180.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "reminder_enabled": {
            "default": true,
            "title": "Reminder Enabled",
            "type": "boolean"
          },
          "source_options": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Source Options",
            "type": "array"
          },
          "starts_at": {
            "format": "date-time",
            "title": "Starts At",
            "type": "string"
          },
          "topic": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic"
          }
        },
        "required": [
          "name",
          "chapter_id",
          "location",
          "starts_at",
          "ends_at",
          "source_options"
        ],
        "title": "EventCreate",
        "type": "object"
      },
      "EventOut": {
        "properties": {
          "archived": {
            "title": "Archived",
            "type": "boolean"
          },
          "attendee_count": {
            "title": "Attendee Count",
            "type": "integer"
          },
          "chapter_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapter Id"
          },
          "chapter_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapter Name"
          },
          "ends_at": {
            "format": "date-time",
            "title": "Ends At",
            "type": "string"
          },
          "feedback_enabled": {
            "title": "Feedback Enabled",
            "type": "boolean"
          },
          "help_options": {
            "items": {
              "type": "string"
            },
            "title": "Help Options",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "locale": {
            "enum": [
              "nl",
              "en"
            ],
            "title": "Locale",
            "type": "string"
          },
          "location": {
            "title": "Location",
            "type": "string"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "reminder_enabled": {
            "title": "Reminder Enabled",
            "type": "boolean"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "source_options": {
            "items": {
              "type": "string"
            },
            "title": "Source Options",
            "type": "array"
          },
          "starts_at": {
            "format": "date-time",
            "title": "Starts At",
            "type": "string"
          },
          "topic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic"
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "topic",
          "location",
          "latitude",
          "longitude",
          "starts_at",
          "ends_at",
          "source_options",
          "help_options",
          "feedback_enabled",
          "reminder_enabled",
          "locale",
          "chapter_id",
          "chapter_name",
          "attendee_count",
          "archived"
        ],
        "title": "EventOut",
        "type": "object"
      },
...
      "FeedbackFormOut": {
        "description": "Public payload for GET /feedback/{token} \u2014 what the questionnaire\npage renders. Includes minimal event context so the page can show\nthe user which event they're giving feedback on.",
        "properties": {
          "event_locale": {
            "enum": [
              "nl",
              "en"
            ],
            "title": "Event Locale",
            "type": "string"
          },
          "event_name": {
            "title": "Event Name",
            "type": "string"
          },
          "event_slug": {
            "title": "Event Slug",
            "type": "string"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/FeedbackQuestionOut"
            },
            "title": "Questions",
            "type": "array"
          }
        },
        "required": [
          "event_name",
          "event_slug",
          "event_locale",
          "questions"
        ],
        "title": "FeedbackFormOut",
        "type": "object"
      },
      "FeedbackQuestionOut": {
        "description": "Question shape for the public form. The frontend resolves the\nprompt + labels from i18n using ``key``. ``key`` is also the\nstable identifier the submit endpoint expects in answers.",
        "properties": {
          "key": {
            "title": "Key",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "ordinal": {
            "title": "Ordinal",
            "type": "integer"
          },
          "required": {
            "title": "Required",
            "type": "boolean"
          }
        },
        "required": [
          "key",
          "ordinal",
          "kind",
          "required"
        ],
        "title": "FeedbackQuestionOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Opkomst",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/events": {
      "get": {
        "operationId": "list_events_api_v1_events_get",
        "parameters": [
          {
            "in": "query",
            "name": "chapter_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Chapter Id"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EventOut"
                  },
                  "title": "Response List Events Api V1 Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Events",
        "tags": [
          "events"
        ]
      },
      "post": {
        "operationId": "create_event_api_v1_events_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Event",
        "tags": [
          "events"
        ]
      }
    },
...
    "/api/v1/events/by-slug/{slug}": {
      "get": {
        "description": "Public event lookup. The response is identical for every\nvisitor in the seconds between two organiser edits, so it's a\nperfect candidate for HTTP caching: a 60 s shared cache window\n(Coolify/Traefik / any CDN that fronts the app) keeps the\ncommon case off the DB entirely, and ``stale-while-revalidate``\nmeans a request after the 60 s expiry serves the still-warm\npayload while a single background fetch refreshes it.\n\nThe 60 s freshness window is the trade-off: an organiser edit\nis visible to *new* visitors after up to 60 s. Acceptable \u2014\nedits during an active sign-up window are rare, and\n``no-store`` would defeat the point of caching the highest-\ntraffic public endpoint.",
        "operationId": "get_event_by_slug_api_v1_events_by_slug__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Event By Slug",
        "tags": [
          "events"
        ]
      }
    },
    "/api/v1/events/by-slug/{slug}/email-preview/{channel}": {
      "get": {
        "description": "Render the exact email that the dispatcher will send to a\nsignup on this event.",
        "operationId": "email_preview_api_v1_events_by_slug__slug__email_preview__channel__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "channel",
            "required": true,
            "schema": {
              "title": "Channel",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Email Preview",
        "tags": [
          "events"
        ]
      }
    },
    "/api/v1/events/by-slug/{slug}/event.ics": {
      "get": {
        "description": "Public RFC 5545 calendar download for one event. Universal \u2014\nGoogle, Apple, Outlook, Proton, Thunderbird, every mobile\ncalendar app imports it. UID is the event's stable ``id``,\nso re-importing after an organiser edit updates the existing\nentry instead of creating a duplicate.",
        "operationId": "get_event_ics_api_v1_events_by_slug__slug__event_ics_get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Event Ics",
        "tags": [
          "events"
        ]
      }
    },
    "/api/v1/events/by-slug/{slug}/feedback-preview": {
      "get": {
        "description": "Preview of the post-event feedback form.",
        "operationId": "feedback_form_preview_api_v1_events_by_slug__slug__feedback_preview_get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackFormOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Feedback Form Preview",
        "tags": [
          "events"
        ]
      }
    },
    "/api/v1/events/by-slug/{slug}/qr.svg": {
      "get": {
        "operationId": "get_event_qr_api_v1_events_by_slug__slug__qr_svg_get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Event Qr",
        "tags": [
          "events"
        ]
      }
    },
...
    "/api/v1/events/{event_id}": {
      "delete": {
        "description": "Hard-delete an archived event. Refuses if the event isn't\narchived first \u2014 accidentally hard-deleting a live event\nwith sign-ups would be a data-loss footgun. Cascades through\n``signups`` / ``email_dispatches`` / ``feedback_responses`` /\n``feedback_tokens`` via the FK ``ON DELETE CASCADE``s in the\nschema; the row + its dependents go with one DELETE.",
        "operationId": "delete_event_api_v1_events__event_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Event",
        "tags": [
          "events"
        ]
      },
      "put": {
        "operationId": "update_event_api_v1_events__event_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Event",
        "tags": [
          "events"
        ]
      }
    },
...
  }
}